Create idapal_automated.py
Note: Modified to add automation. Performs analysis on all well defined functions. Ordering of function examination follows a relatively simple method based on the number of call instructions (excluding imports). This was done this way in the idea it would add better information later on in the examination to the larger functions. This expected behavior has not yet been validated.

Best,
-Me